Update Using Environments to explain per-environment API keys (#722)
This improves the documentation regarding per-environment API keys. As a user of Statsig, I was initially confused about why these keys were segmented per environment.

My first thoughts were that it was going to affect how the evaluation context was used (similar to how LaunchDarkly's evaluation works when using different environment SDK keys). This isn't the case with Statsig, it's more of a measure to restrict what data each SDK has access to.