add id field to rule (#1405)
Add 'id' field to rule, both in the request body and the response body. Also added 'environments' field as well, looks like it had yet to be added to the docs.

NOTE: the current system of using 1 rule model across both requests and responses for all of gates/segments/dynamic configs is not quite right, there should be some redesign here. e.g., the 'returnValue' JSON field is a required field for gate updates/responses, although this is not true of the actual API