Add environment overrides to holdout docs (#1622)
Holdouts support environment-based overrides, the same as gates do. A customer ran into some confusion around the lack of documentation for the "environment" field on a holdout override.

This PR updates the openapi documentation to include the optional "environmentOverrides" field, copying over the ref type for "environment_overrides" (tweaking some wording to be holdout-specific).