help make finding code pointers easy (#3190)
## Description

Adding a description of where to find code pointers on the feature flagging, dynamic config, and experiment guides in our docs.

Also adding a link to the SDK section for dynamic configs to hopefully help ppl head in the right direction if they want SDK help but end up on the tutorials/guides on the product section of our docs.